CalendarObject as a Readable stream
This PR addresses tritech/node-icalendar/issues/29 see that issue for
the reasoning.

Here are some notes from implementing.

It wasn’t in my scope to allow the events themselves to be streams as
well.  Currently only the calendar object is streamable, however as you
can see in the tests you can add events to that object and the whole is
correctly streamed.

So for example this is currently not possible:

```javascript
var vevent = new icalendar.VEvent('testuid@example.com');
vevent.stream(s3.uploadStream('ical', 'uid.ics’));
```

Yet you can easily create a calendar, add the event to that calendar
and stream as expected above.

will need to require node 0.10 or greater to use the streams2 API